@@ -1,18 +1,19 @@
 {
   "name": "js-message",
-  "version": "1.0.7",
-  "description": "normalized JS Object and JSON message and event protocol for node.js, vanialla js, react.js, components, actions, stores and dispatchers",
-  "main": "Message.js",
+  "version": "2.0.0",
+  "description": "normalized JS Object and JSON message and event protocol for ES6+ node.js, browsers, electron, vanialla js, react.js, components, actions, stores and dispatchers",
+  "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node devServer.js"
+    "start": "node-http-server port=8000 verbose=true"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/RIAEvangelist/js-message.git"
   },
   "engines": {
-    "node": ">=0.6.0"
+    "node": ">=13.2.0"
   },
   "keywords": [
     "message",
@@ -50,5 +51,8 @@
   },
   "homepage": "https://github.com/RIAEvangelist/js-message#readme",
   "dependencies": {},
-  "devDependencies": {}
+  "devDependencies": {
+    "node-http-server": "^8.1.3",
+    "vanilla-test": "^1.4.2"
+  }
 }
